Refactorings and improvements for CMEMS and GFS importers
