+++ docker.env
@@ -1,7 +1,7 @@
-DB_HOST=db
-DB_USER=geonode-postgres-admin
-DB_PASSWORD=geonode-postgres-admin-passwort
-DB=geoserver
+DB_HOST=database
+DB_USER=opendatacube
+DB_PASSWORD=opendatacube
+DB=opendatacube
 NO_PING=True
 SLEEP=2
 MAX_RETRIES=15
